+++ packages/cli/src/index.ts
@@ -10,6 +10,7 @@ import { registerFingerprint } from './commands/fingerprint.js'
 import { registerSelfCheck } from './commands/self-check.js'
 import { registerTest } from './commands/test.js'
 import { registerCapabilities } from './commands/capabilities.js'
+import { registerPolicyDiff } from './commands/policy-diff.js'
 
 const program = new Command()
 
@@ -26,5 +27,6 @@ registerFingerprint(program)
 registerSelfCheck(program)
 registerTest(program)
 registerCapabilities(program)
+registerPolicyDiff(program)
 
 program.parse(process.argv)